Separate boxes and flow from normal text layout as the code was getting very large for one file with the new changes. Clean up GlyphBox.place into separate functions for the different kinds of text. Only clear cached groups when group or program changes now.
